refactor: remove custom components
