--- src/components/atoms/card-content.vue
@@ -1,9 +1,0 @@
-<template>
-  <q-card flat bordered>
-    <slot></slot>
-  </q-card>
-</template>
-
-<script setup lang="ts"></script>
-
-<style scoped></style>
