+++ src/components/atoms/index.ts
@@ -1,3 +1,1 @@
-export { default as AppButton } from './app-button.vue'
-export { default as CardContent } from './card-content.vue'
-export { default as PageTitle } from './page-title.vue'
+//export { default as Component } from './component.vue'
