+++ src/components/index.ts
@@ -1,3 +1,3 @@
-export * from './atoms'
-export * from './molecules'
-export * from './organisms'
+//export * from './atoms'
+//export * from './molecules'
+//export * from './organisms'
