+++ src/components/molecules/index.ts
@@ -1,3 +1,1 @@
-export { default as DataTable } from './data-table.vue'
-export { default as MemberCard } from './member-card.vue'
-export { default as SidebarMenu } from './sidebar-menu.vue'
+//export { default as Component } from './component.vue'
