--- src/components/organisms/requests-board.vue
@@ -1,7 +1,0 @@
-<template>
-  <div>Requests</div>
-</template>
-
-<script setup lang="ts"></script>
-
-<style scoped></style>
