+++ src/core/locales/en.json
@@ -1,10 +1,9 @@
 {
   "common": {
     "app": {
-      "name": "HRMS",
+      "name": "Vue Base Enterprise v1.0",
       "version": "ver.1.0.0",
-      "sidebar_default_menu": "Main Menu",
-      "sidebar_admin_menu": "Administration"
+      "sidebar_default_menu": "Main Menu"
     },
     "buttons": {
       "submit": "Submit",
