+++ src/core/types/app-config.type.ts
@@ -1,4 +1,5 @@
 export type AppConfig = {
+  loadCoreUiComponents: () => AppConfig
   loadComponents: () => AppConfig
   setLocal: (locale: string) => AppConfig
   setStore: () => AppConfig
