+++ src/core/ui/atoms/index.ts
@@ -1,0 +1,2 @@
+export { default as AppButton } from './app-button.vue'
+export { default as PageTitle } from './page-title.vue'
