+++ src/core/ui/molecules/index.ts
@@ -1,0 +1,2 @@
+export { default as DataTable } from './data-table.vue'
+export { default as SidebarMenu } from './sidebar-menu.vue'
