+++ src/core/ui/molecules/sidebar-menu.vue
