+++ src/core/ui/organisms/index.ts
@@ -1,0 +1,2 @@
+export { default as AppDrawer } from './app-drawer.vue'
+export { default as AppHeader } from './app-header.vue'
