--- src/pages/administration/dashboard/index-page.vue
@@ -1,7 +1,0 @@
-<template>
-  <div>Dashboard</div>
-</template>
-
-<script setup lang="ts"></script>
-
-<style scoped></style>
