--- src/pages/administration/members/index-page.vue
@@ -1,7 +1,0 @@
-<template>
-  <MemberBoard></MemberBoard>
-</template>
-
-<script setup lang="ts"></script>
-
-<style scoped></style>
