--- src/pages/administration/requests/index-page.vue
@@ -1,7 +1,0 @@
-<template>
-  <RequestsBoard></RequestsBoard>
-</template>
-
-<script setup lang="ts"></script>
-
-<style scoped></style>
