+++ src/pages/index-page.vue
@@ -1,9 +1,6 @@
 <template>
-
   <div class="row">
-    <div class="col-12 col-sm-12 col-md-6 col-lg-4">
-      <DashboardTimelog></DashboardTimelog>
-    </div>
+    <div class="col-12 col-sm-12 col-md-6 col-lg-4">Home Page</div>
   </div>
 </template>
