--- src/pages/leaves/index-page.vue
@@ -1,7 +1,0 @@
-<template>
-  <div>Leaves</div>
-</template>
-
-<script setup lang="ts"></script>
-
-<style scoped></style>
