--- src/pages/payslips/index-page.vue
@@ -1,7 +1,0 @@
-<template>
-  <div>Paylips</div>
-</template>
-
-<script setup lang="ts"></script>
-
-<style scoped></style>
