--- src/pages/profile/index-page.vue
@@ -1,7 +1,0 @@
-<template>
-  <div>Profile</div>
-</template>
-
-<script setup lang="ts"></script>
-
-<style scoped></style>
