--- src/pages/timesheet/index-page.vue
@@ -1,11 +1,0 @@
-<template>
-  <DataTable :columns="columns" :data="timelogs" />
-</template>
-
-<script setup lang="ts">
-import { useTimesheet } from 'src/core/composables'
-
-const { columns, timelogs } = useTimesheet()
-</script>
-
-<style scoped></style>
